Publish locally produced nuget package for easy consumption
Locally testing nuget-producing projects is always a bit tricky.
You have to basically:

1 - Clean the output folder from previously produced packages
2 - Add the output folder as a nuget source on the consuming project
3 - Clean the nuget restore cache so that the newly produced .nupkg
    from the new source will always be unpacked, rather than picked
    up from a previous build already cached
4 - Repeat for every project that produces packages ;-)

We can do better!

This change adds a target `_LocalPublish` that will only run on local
builds, which does all of the above, but copying the output to
`%TEMP%\packages`, so that to test packages, you only need to add a single
source in a `NuGet.Config` (could even be a machine-wide one ;)):

```xml
<configuration>
	<packageSources>
		<add key="local" value="%temp%\packages" />
	</packageSources>
</configuration>
```

Re-packing the project would automatically clear the nuget cache (just
for the package machine the current `$(PackageId)` of course), delete
previously built versions, and copy the output to the temp path.

This way, rapidly iterating on the packaging output becomes even
enjoyable!