Make sure VS builds when we change targets
By default, VS will not consider None files that are not copied to
output to have any effect on build, and therefore changing targets
will result in VS saying there's nothing to build.

In order for the Build/Pack to trigger anytime we make changes in
any targets file, add the @(None) to the up-to-date check input
files.

Found via https://github.com/dotnet/project-system/pull/2416