Delete local nuget cache at the end
Ensure the copying of the new version is performed before attempting
to delete the user cache, since VS might be holding some files or
folders and delete sometimes fails, but the new package can still
work fine. Since this is sort of random, I'd rather have the build
fail *after* the local package was copied to $temp$\packages, at
which point I can try out myself to see if the deletion of the
cache is actually required or not for local testing.