Add support for smart platform defaults and nesting in the IDE
Provides default items per-platform so that authoring multi-targeting projects
is as straightforward as placing the source files within platform-specific folders
(such as Platforms/Android/Renderer.cs or Android/Renderer.cs) as well as the more
lightweight option of just adding a target platform suffix to source files (such
as FancyButton.Android.cs and FancyButton.Tizen.cs), which additionally get
automatically nested (%(DependentUpon) is properly set) to the "parent" source
file (such as FancyButton.cs).

The right target frameworks are automatically set to compile the given items
as long as they match the above criteria for either:

* full target framework identifier in path (i.e. `NETFramework` or `Xamarin.iOS`)
* target framework identifier plus version (i.e. `NETFramework4.7` or `Xamarin.Android8.0`),
  which allows version-specific switching (for cases where you want to tweak the code
  for, say, `monoandroid8.0` vs `monoandroid7.0`)
* target framework identifier plus major version (i.e. `NETStandard2`)
* short framework identifier (i.e. `monoandroid`)
* short framework identifier plus version (i.e. `monoandroid8.0`)
* short framework identifier plus major version (i.e. `monoandroid8` vs `monoandroid7`,
  or `netstandard1` vs `netstandard2` if you expose different implementations)

Additionally, the following frameworks are given short aliases that are commonly
used in projects:

* Xamarin.Android => Android (i.e. place stuff under `Platforms/Android/*.cs` or
  `Android/*.cs` or name a file `Foo.Android.cs`)
* Xamarin.iOS => iOS
* Xamarin.Mac => Mac
* Xamarin.TVOS => TVOS
* Xamarin.WatchOS => WatchOS

Also, `NETStandard` and `netstandard` can be used interchangeable as long as the
volume where MSBuild is running has not been created as a case-sensitive volume
(HFS+ and APFS are both case insensitive by default). Otherwise, the files on
disk must match the `NETStandard` casing since that's the proper Target Framework
Identifier for netstandard.

A sample/test `ControlLibrary.csproj` project is provided that showcases the various
use cases.

NOTE: the project system is put under some pressure from all these wildcards, so
adding/renaming items is slow-ish, but otherwise, it all just works beautifully.
In case the performance deteriorates too much and the feature is not used, the
`$(EnableDefaultPlatformItems)` can be set to `false` to disable the feature
entirely if not used.