Add and use a NuGet package in the WPF sample
This makes failure to pick up the NuGet packages
an easy-to-detect compile error. I used PCLCommandBase
because that was used in
https://github.com/NuGet/Home/issues/4652#issuecomment-280930062

Furthermore, note that I did not retarget the sample project
to net472, yet it still works. net472 must be *installed* on
the build machine for this change to succeed; it need not be
*targeted*.