Use $(_TargetAssemblyProjectName) in Wpf.targets (#68)
* Add latest WPF/NuGet workaround

See https://github.com/NuGet/Home/issues/5894#issuecomment-388520580 for context.

* Add %TEMP%\packages to NuGet.config

This ensures that the local dev packages automatically
copied into that directory by MSBuild.Sdk.Extras can be
used by sample and test projects.

* Rewrite WpfApplication sample to consume SDK

* Add and use a NuGet package in the WPF sample

This makes failure to pick up the NuGet packages
an easy-to-detect compile error. I used PCLCommandBase
because that was used in
https://github.com/NuGet/Home/issues/4652#issuecomment-280930062

Furthermore, note that I did not retarget the sample project
to net472, yet it still works. net472 must be *installed* on
the build machine for this change to succeed; it need not be
*targeted*.

* Move WpfAppUsingSdk into tests directory per request

* Remove local key from top-level NuGet.config

* Move project into correct solution

* Move WPF NuGet props import

Also updated a few comments while I was at it.